// gravitee-apim-portal-webui/projects/portal-webclient-sdk/src/lib/model/filterApiQuery.ts
/**
 * Gravitee.io Portal Rest API
 * API dedicated to the devportal part of Gravitee
 *
 * The version of the OpenAPI document: 3.14.0-SNAPSHOT
 * Contact: contact@graviteesource.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export type FilterApiQuery = 'ALL' | 'FEATURED' | 'MINE' | 'STARRED' | 'TRENDINGS';

export const FilterApiQuery = {
    ALL: 'ALL' as FilterApiQuery,
    FEATURED: 'FEATURED' as FilterApiQuery,
    MINE: 'MINE' as FilterApiQuery,
    STARRED: 'STARRED' as FilterApiQuery,
    TRENDINGS: 'TRENDINGS' as FilterApiQuery
};